POText.measuredPaddingTop() applied to PORadioGroup
